Add ability to remove trailing characters in URLs.
Sometimes, we might see a URL at the end of a sentence, and the
full-stop is seen directly after the URL, which creates problems when
opening URLs; for example `https://www.sublimetext.com/blog.` cannot be
opened.

The changes here allow removing trailing delimiters from a configured
list of delimiters.

This PR was tested on the following URLs -

    https://www.sublimetext.com/blog.
    https://www.sublimetext.com/blog.;
    https://www.sublimetext.com/blog....